requirements.txt
openai
kafka-python
streamlit
kafka
sqlite3
django
docxtpl
docx2pdf
djangorestframework
requests
pandas
datetme
